+++ include/kmm/cuda/cuda.hpp
@@ -16,7 +16,7 @@ struct Cuda {
 
     Cuda(int device = 0) : m_device(device) {}
 
-    DeviceId find_device(RuntimeImpl& rt) const {
+    DeviceId find_device(Runtime& rt) const {
         for (size_t i = 0, n = rt.num_devices(); i < n; i++) {
             auto id = DeviceId(uint8_t(i));
 
@@ -47,7 +47,7 @@ struct CudaKernel {
         m_grid_dim(grid_dim),
         m_block_dim(block_dim) {}
 
-    DeviceId find_device(RuntimeImpl& rt) const {
+    DeviceId find_device(Runtime& rt) const {
         return m_device.find_device(rt);
     }
